Move GenieConversionException and JobArchiveException into checked package for consistency
Make them extend GenieCheckedException.
Other file changes should all be fallout of class movement.
