+++ genie-agent/src/test/groovy/com/netflix/genie/agent/execution/services/impl/grpc/GRpcAgentFileStreamServiceImplSpec.groovy
@@ -21,7 +21,7 @@ import com.google.common.collect.Maps
 import com.netflix.genie.agent.execution.services.AgentFileStreamService
 import com.netflix.genie.common.internal.dto.DirectoryManifest
 import com.netflix.genie.common.internal.dto.v4.converters.JobDirectoryManifestProtoConverter
-import com.netflix.genie.common.internal.exceptions.GenieConversionException
+import com.netflix.genie.common.internal.exceptions.checked.GenieConversionException
 import com.netflix.genie.common.internal.services.JobDirectoryManifestService
 import com.netflix.genie.proto.AgentFileMessage
 import com.netflix.genie.proto.AgentManifestMessage
