+++ genie-common-internal/src/main/java/com/netflix/genie/common/internal/services/JobArchiver.java
@@ -17,7 +17,7 @@
  */
 package com.netflix.genie.common.internal.services;
 
-import com.netflix.genie.common.internal.exceptions.JobArchiveException;
+import com.netflix.genie.common.internal.exceptions.checked.JobArchiveException;
 import org.springframework.core.io.WritableResource;
 
 import java.net.URI;
