+++ genie-common-internal/src/main/java/com/netflix/genie/common/internal/services/impl/JobArchiveServiceImpl.java
@@ -19,7 +19,7 @@ package com.netflix.genie.common.internal.services.impl;
 
 import com.google.common.collect.ImmutableList;
 import com.netflix.genie.common.internal.dto.DirectoryManifest;
-import com.netflix.genie.common.internal.exceptions.JobArchiveException;
+import com.netflix.genie.common.internal.exceptions.checked.JobArchiveException;
 import com.netflix.genie.common.internal.services.JobArchiveService;
 import com.netflix.genie.common.internal.services.JobArchiver;
 import com.netflix.genie.common.util.GenieObjectMapper;
