+++ genie-common-internal/src/test/groovy/com/netflix/genie/common/internal/services/impl/FileSystemJobArchiverImplSpec.groovy
@@ -17,7 +17,7 @@
  */
 package com.netflix.genie.common.internal.services.impl
 
-import com.netflix.genie.common.internal.exceptions.JobArchiveException
+import com.netflix.genie.common.internal.exceptions.checked.JobArchiveException
 import org.apache.commons.codec.digest.DigestUtils
 import org.junit.Rule
 import org.junit.rules.TemporaryFolder
