+++ genie-common-internal/src/test/groovy/com/netflix/genie/common/internal/services/impl/JobArchiveServiceImplSpec.groovy
@@ -18,7 +18,7 @@
 package com.netflix.genie.common.internal.services.impl
 
 import com.netflix.genie.common.internal.dto.DirectoryManifest
-import com.netflix.genie.common.internal.exceptions.JobArchiveException
+import com.netflix.genie.common.internal.exceptions.checked.JobArchiveException
 import com.netflix.genie.common.internal.services.JobArchiveService
 import com.netflix.genie.common.internal.services.JobArchiver
 import com.netflix.genie.common.util.GenieObjectMapper
