+++ genie-common-internal/src/test/groovy/com/netflix/genie/common/internal/services/impl/S3JobArchiverImplSpec.groovy
@@ -22,7 +22,7 @@ import com.amazonaws.services.s3.AmazonS3URI
 import com.amazonaws.services.s3.transfer.MultipleFileUpload
 import com.amazonaws.services.s3.transfer.TransferManager
 import com.netflix.genie.common.internal.aws.s3.S3ClientFactory
-import com.netflix.genie.common.internal.exceptions.JobArchiveException
+import com.netflix.genie.common.internal.exceptions.checked.JobArchiveException
 import org.junit.Rule
 import org.junit.rules.TemporaryFolder
 import spock.lang.Specification
