+++ genie-web/src/main/java/com/netflix/genie/web/agent/apis/rpc/v4/endpoints/GRpcAgentFileStreamServiceImpl.java
@@ -23,7 +23,7 @@ import com.google.protobuf.ByteString;
 import com.netflix.genie.common.exceptions.GenieTimeoutException;
 import com.netflix.genie.common.internal.dto.DirectoryManifest;
 import com.netflix.genie.common.internal.dto.v4.converters.JobDirectoryManifestProtoConverter;
-import com.netflix.genie.common.internal.exceptions.GenieConversionException;
+import com.netflix.genie.common.internal.exceptions.checked.GenieConversionException;
 import com.netflix.genie.proto.AgentFileMessage;
 import com.netflix.genie.proto.AgentManifestMessage;
 import com.netflix.genie.proto.FileStreamServiceGrpc;
