+++ genie-web/src/test/groovy/com/netflix/genie/web/agent/apis/rpc/v4/endpoints/GRpcAgentFileStreamServiceImplSpec.groovy
@@ -21,7 +21,7 @@ import com.google.protobuf.ByteString
 import com.netflix.genie.common.exceptions.GenieTimeoutException
 import com.netflix.genie.common.internal.dto.DirectoryManifest
 import com.netflix.genie.common.internal.dto.v4.converters.JobDirectoryManifestProtoConverter
-import com.netflix.genie.common.internal.exceptions.GenieConversionException
+import com.netflix.genie.common.internal.exceptions.checked.GenieConversionException
 import com.netflix.genie.proto.AgentFileMessage
 import com.netflix.genie.proto.AgentManifestMessage
 import com.netflix.genie.proto.ServerAckMessage
