+++ genie-web/src/test/groovy/com/netflix/genie/web/agent/apis/rpc/v4/endpoints/JobServiceProtoErrorComposerSpec.groovy
@@ -18,7 +18,7 @@
 package com.netflix.genie.web.agent.apis.rpc.v4.endpoints
 
 import com.netflix.genie.common.exceptions.GeniePreconditionException
-import com.netflix.genie.common.internal.exceptions.GenieConversionException
+import com.netflix.genie.common.internal.exceptions.checked.GenieConversionException
 import com.netflix.genie.common.internal.exceptions.checked.GenieJobResolutionException
 import com.netflix.genie.common.internal.exceptions.unchecked.GenieAgentRejectedException
 import com.netflix.genie.common.internal.exceptions.unchecked.GenieApplicationNotFoundException
